Moved everything to the inside of bucket package
