+++ bucket/protocol.go
@@ -1,2 +1,2 @@
 //go:generate protoc --gogoslick_out=. protocol.proto
-package kadiradb
+package bucket
